+++ backend/src/main/java/com/company/evaluation/employee/EmployeeController.java
@@ -21,9 +21,11 @@ import java.util.List;
 @CrossOrigin
 public class EmployeeController {
     private final EmployeeService service;
+    private final EmployeeExcelService excelService;
 
-    public EmployeeController(EmployeeService service) {
+    public EmployeeController(EmployeeService service, EmployeeExcelService excelService) {
         this.service = service;
+        this.excelService = excelService;
     }
 
     @GetMapping
@@ -44,48 +46,7 @@ public class EmployeeController {
     // 사번 | 사원명 | 입사년도 | 급여 | 직급 | 근무상태 | 근무지역 | 이메일 | 전화번호 | 메모
     @PostMapping(value = "/upload", consumes = {"multipart/form-data"})
     public ApiResponse<Integer> upload(@RequestPart("file") MultipartFile file) throws Exception {
-        int saved = 0;
-        try (InputStream is = file.getInputStream(); Workbook wb = new XSSFWorkbook(is)) {
-            Sheet sheet = wb.getSheetAt(0);
-            List<EmployeeRequest> rows = new ArrayList<>();
-            boolean header = true;
-            for (Row row : sheet) {
-                if (header) { header = false; continue; }
-                if (row == null) continue;
-                EmployeeRequest dto = new EmployeeRequest();
-                dto.setEmployeeNumber(getString(row, 0));
-                dto.setName(getString(row, 1));
-                dto.setJoinYear(getInt(row, 2));
-                dto.setBaseSalary(getLong(row, 3));
-                dto.setPosition(getString(row, 4));
-                dto.setWorkStatus(getString(row, 5));
-                dto.setWorkRegion(getString(row, 6));
-                dto.setEmail(getString(row, 7));
-                dto.setPhone(getString(row, 8));
-                dto.setNotes(getString(row, 9));
-                if (dto.getEmployeeNumber() != null && !dto.getEmployeeNumber().isBlank()) {
-                    rows.add(dto);
-                }
-            }
-            for (EmployeeRequest dto : rows) {
-                service.save(dto);
-                saved++;
-            }
-        }
-        return ApiResponse.ok(saved);
-    }
-
-    private static String getString(Row row, int idx) {
-        if (row.getCell(idx) == null) return null;
-        row.getCell(idx).setCellType(org.apache.poi.ss.usermodel.CellType.STRING);
-        String s = row.getCell(idx).getStringCellValue();
-        return s != null ? s.trim() : null;
-    }
-    private static Integer getInt(Row row, int idx) {
-        try { return (int) Math.round(row.getCell(idx).getNumericCellValue()); } catch (Exception e) { return null; }
-    }
-    private static Long getLong(Row row, int idx) {
-        try { return (long) Math.round(row.getCell(idx).getNumericCellValue()); } catch (Exception e) { return null; }
+        return ApiResponse.ok(excelService.importExcel(file));
     }
